Added basiic ballista controls
ballista cant rotate left to right yet but has up and down and bolt gives visual feedback as to how far it has been pulled back. Ballista control might be better down with the stick and simply have the bolt control with the controller. Or might just need to tweak values for rotation of the ballista